Set pendingWakeup in IRQfindNextThread() instead of only in IRQtimerInterrupt()
This is important as IRQfindNextThread() is also called by peripheral
drivers, and pendingWakeup is important to make sure context switches to
a higher priority thread happen in a timely fashion. Even more
important, it prevents a deadlock when there is a single thread in the
system and the peripheral interrupt waking the thread occurs precisely
while the idle thread is in a PauseKernelLock.